+++ NEWS.md
@@ -3,6 +3,7 @@
 * standardised file names
 * much improved documentation
 * more coherent `makeCPO*` functions
+* %<>>% and %<<<% operators for compose / attach + assignment 
 
 # mlrCPO 0.1
 * Initial release on GitHub, retrafoless and target operating CPOs don't work yet